refactor(lens): split lens.py into per-domain registration package
lens.py grew to 2071 lines with 65 ToolDefs from six separate feature
epics. Every future mutating tool (#1282 has 6 more coming) would grow
it further. Split into a package so each domain gets its own file and
new tools land in the right place instead of tail-appending.

Structure — apps/api/app/tools/registrations/lens/
  __init__.py            composition root: imports submodules + register()
  _shared.py             constants, _impl, _actor_impl, _window_start
  guard_core.py          spend, events, sessions, memory, knowledge, budgets, savings, correlated events, config (13 tools)
  policies.py            list/get_policy, list_credentials (3 tools)
  workflows.py           list_runs, get_run, list_run_events, list_workflows, playbooks (8 tools)
  workspace.py           projects, members, integrations, agent identities, workspace KPIs (9 tools)
  marketplace.py         installed packs, browse, pack details (3 tools)
  ops.py                 pending approvals, alerts, audit log, autopilot (7 tools)
  discovery.py           discovery summary + discovered agents (2 tools)
  primitives.py          LLM primitives, sync state, rate limits (3 tools)
  governance.py          framework coverage, SOC2, AI rollout, compliance status, governance rollups (8 tools)
  dashboard_kpis.py      #1439 dashboard batch (5 tools)
  observability_kpis.py  #1439 observability batch (5 tools)
  actor.py               #1282 mutating tools placeholder

Loading contract preserved
- No behavior change; default_registry ends up populated with the same
  65 ToolDefs after import.
- Each submodule imports only from _shared — no cross-submodule imports,
  no cycles, deterministic import order.
- Backwards-compat re-exports in __init__.py preserve every public symbol
  external callers (mostly tests) used to import from the flat file.

Tests
- Full apps/api tests/tools + tests/glens suite green (99 passed, 18 skipped, 0 failures).
- Only test change: test_governance_rollup_tools.py — 3 patch targets moved
  from `lens._run_framework_coverage` to `lens.governance._run_framework_coverage`
  since the helper now lives in the domain file.

File sizes after split
- Largest: observability_kpis.py 376, dashboard_kpis.py 335. All others under 240.
- No file over 400 lines. All searchable by domain.

Adding a new tool now
- Read tool: find the matching domain file, append a ToolDef to its TOOLS list.
- Mutating tool: register an ActionSpec under app.modules.glens.actor.registrations
  and add a paired ToolDef to lens/actor.py with impl=_actor_impl("name").
